rename stove.js to sf-stove.js
